Add files via upload
An academic group project completed for PSTAT 175 under the supervision of Professor Andrew Carter at UCSB. My group was interested in seeing how the different covariates affect the survival probability of the couple’s marriages, and we aimed to examine the interactions between the covariates to see if any of these interactions are significant. Furthermore, we were also interested in seeing whether or not the covariates have differing effects on the probability of divorce depending on how long a couple has been married. We initially hypothesized that the hazard rate is high at the beginning of a marriage, and that, after a certain point in time, the hazard rate will decrease significantly. Our ultimate goal for this project is to test whether or not this hypothesis was true by using various statistical tools to successfully manipulate and analyze our data.